+++ Team41/Code41/src/spa/src/PKB/OneToMany.h
@@ -103,14 +103,14 @@ public:
      *
      * @return number of keys
      */
-    int keySize();
+    size_t keySize();
 
     /**
      * Gets the number of values stored in table
      *
      * @return number of vals
      */
-    int valSize();
+    size_t valSize();
 };
 
 #include "OneToMany.tpp"
